+++ HISTORY.md
@@ -4,6 +4,7 @@ unreleased
   * Attempt to combine multiple ranges into single range
   * Correctly inherit from `Stream` class
   * Fix `Content-Range` header in 416 responses when using `start`/`end` options
+  * Fix `Content-Range` header missing from default 416 responses
   * Ignore non-byte `Range` headers
   * deps: http-errors@~1.5.0
     - Add `HttpError` export, for `err instanceof createError.HttpError`